Reorganize filing structure a little to have a component folder. Add Navigation, and routes.
