+++ src/App.css
@@ -2,14 +2,8 @@
   text-align: center;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
 .App-header {
   background-color: #222;
-  height: 150px;
   padding: 20px;
   color: white;
 }
@@ -22,7 +16,34 @@
   font-size: large;
 }
 
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
+.navigation {
+  background-color: #222;
+  display: flex;
+  justify-content: center;
+  padding-bottom: 1rem;
+}
+
+.navigation-list {
+  display: flex;
+  flex-direction: row;
+  list-style: none;
+}
+
+.list-item {
+  margin-right: 1rem;
+  text-decoration: none;
+  color: black;
+  background-color: green;
+  padding: 0.5rem;
+  border-radius: 3px;
+  transition: padding 0.1s;
+}
+
+.list-item:hover {
+  margin-right: 1rem;
+  text-decoration: none;
+  color: black;
+  background-color: blue;
+  padding: 0.6rem;
+  border-radius: 3px;
 }
