+++ src/components/App.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
-import './App.css';
-import base from './base.js';
+import '../App.css';
+import base from '../base.js';
 import Plant from './Plant.js';
+import { BrowserRouter as Router } from 'react-router-dom';
+import Navigation from './Navigation.js';
 
 class App extends Component {
   constructor(props) {
@@ -34,8 +36,11 @@ class App extends Component {
     return (
       <div className="App">
         <header className="App-header">
-          <h1 className="App-title">Welcome to React</h1>
+          <h1 className="App-title">Welcome to Water Me</h1>
         </header>
+        <Router>
+          <Navigation />
+        </Router>
         <p className="App-intro">
           To get started, edit <code>src/App.js</code> and save to reload.
         </p>
