+++ src/components/Plant.js
