+++ src/constants/routes.js
@@ -1,0 +1,6 @@
+export const SIGN_UP = '/signup';
+export const SIGN_IN = '/signin';
+export const LANDING = '/';
+export const HOME = '/home'; //protected, can only be accessed when signed in
+export const ACCOUNT = '/account';
+export const PASSWORD_FORGET = '/pw-forget';
